+++ iOS/G3MiOSSDK/Commons/Rendererers/Renderer.hpp
@@ -52,7 +52,8 @@ public:
   
   virtual PlanetRenderer* getPlanetRenderer() = 0;
     
-  virtual void setChangedRendererInfoListener(ChangedRendererInfoListener* changedInfoListener, const int rendererIdentifier) = 0;
+  virtual void setChangedRendererInfoListener(ChangedRendererInfoListener* changedInfoListener,
+                                              const size_t rendererIdentifier) = 0;
   
 //  virtual void setInfo(const std::vector<std::string>& info) = 0;
